requirements.txt
asgiref==3.3.1
azure-cognitiveservices-vision-customvision==3.1.0
azure-common==1.1.26
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.24.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.6
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.0.4
idna==2.10
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.0
Pillow==8.1.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
urllib3==1.26.3
whitenoise==5.2.0
